% Auto-generated: do not edit by hand
\name{dashTabulator}

\alias{dashTabulator}

\title{DashTabulator component}

\description{

}

\usage{
dashTabulator(id=NULL, cellEdited=NULL, columns=NULL, data=NULL,
groupBy=NULL)
}

\arguments{
\item{id}{Character. ID used to identify this component in Dash callbacks.}

\item{cellEdited}{Named list. cellEdited captures the cell that was clicked on}

\item{columns}{Unnamed list. Column definitions.}

\item{data}{Unnamed list. Data to display in the table.}

\item{groupBy}{Unnamed list. Columns to group by.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

