% Auto-generated: do not edit by hand
\name{dashTabulator}

\alias{dashTabulator}

\title{DashTabulator component}

\description{

}

\usage{
dashTabulator(id=NULL, cellEdited=NULL, columns=NULL, data=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{cellEdited}{Named list. cellEdited captures the cell that was clicked on}

\item{columns}{Unnamed list. A label that will be printed when this component is rendered.}

\item{data}{Unnamed list. The value displayed in the input.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

